Time: 2 ms (60.64%), Space: 55.4 MB (84.66%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cs
@@ -50,7 +50,7 @@ public class Solution {
             MID = res[RL / 2 - 1];
             MID2 = res[RL / 2];
 
-            MID = (MID + MID2) / 2;
+            MID = (float)(MID + MID2) / 2;
         }
         else{
